fix: remove mapping file and export subgraphID from handleLiquidatePosition, refactor file structure
