--- src/mapping.ts
@@ -1,5 +1,0 @@
-// // Note: EPNS only supports The Graph Hosted Service at present
-// import { sendEPNSNotification } from './EPNSNotification';
-// import { PositionLiquidation } from './templates/MarginEngine/MarginEngine';
-
-export const subgraphID = 'voltzprotocol/voltz-goerli'; // change to mainnet when deploying to mainnet subgraph
